update contact form
- remove disabled state on submit button
- reorganize action code
- remove excessive client side validation
- use `user-invalid` instead of `invalid` for client side validation
