+++ src/actions/index.ts
@@ -1,16 +1,5 @@
-import sendEmail from "@utils/contact";
-import { defineAction } from "astro:actions";
-import { z } from "astro:schema";
+import { contact } from "./contact";
 
 export const server = {
-  contact: defineAction({
-    accept: "form",
-    input: z.object({
-      name: z.string(),
-      email: z.string().email(),
-      message: z.string(),
-      "cf-turnstile-response": z.string(),
-    }),
-    handler: sendEmail,
-  }),
+  contact,
 };
